Migration from custom int types to standart int

+++ Types and data structures/Lab2/common.c
@@ -1,7 +1,7 @@
 #include "common.h"
 
 //initialisation finctions
-Data_t initGeneratorData(uint16_t width, uint16_t height, Cell_t startPoint){
+Data_t initGeneratorData(int width, int height, Cell_t startPoint){
     Data_t d;
     d.width        = width;
     d.height       = height;
@@ -31,24 +31,24 @@ Data_t initSeekerData(Data_t d, Cell_t startPoint, Cell_t exitPoint){
 }
 
 //generator and seeker common functions
-CellString_t getNeighbours(Data_t *data, Cell_t c, int8_t distance){
-    uint16_t x   = c.x;
-    uint16_t y   = c.y;
+CellString_t getNeighbours(Data_t *data, Cell_t c, int distance){
+    int x   = c.x;
+    int y   = c.y;
     Cell_t up    = {x, y - distance};
     Cell_t rt    = {x + distance, y};
     Cell_t dw    = {x, y + distance};
     Cell_t lt    = {x - distance, y};
     Cell_t d[4]  = {dw, rt, up, lt};
-    uint8_t size = 0;
+    int size = 0;
 
     CellString_t cells;
     cells.cells = malloc(4 * sizeof(Cell_t));
 
-    uint16_t     height = data->height;
-    uint16_t     width  = data->width;
+    int     height = data->height;
+    int     width  = data->width;
     MazeMatrix_t maze   = data->maze;
 
-    for(uint8_t i = 0; i < 4; i++){ //для каждого направдения d
+    for(int i = 0; i < 4; i++){ //для каждого направдения d
         if(d[i].x > 0 && d[i].x < width && d[i].y > 0 && d[i].y < height){ //если не выходит за границ
             CellType_t mazeCellCurrent = maze[d[i].y][d[i].x];
             Cell_t     cellCurrent     = d[i];
@@ -58,19 +58,19 @@ CellString_t getNeighbours(Data_t *data, Cell_t c, int8_t distance){
             }
         }
     }
-    
+
     cells.size = size;
     return cells;
 }
 
 
-uint32_t randomRange(uint32_t low, uint32_t high){
+int randomRange(int low, int high){
     return rand() % (high - low + 1) + low;
 }
 
 
 void setMode(Cell_t c, MazeMatrix_t maze, CellType_t mode){
-    uint16_t x = c.x;
-    uint16_t y = c.y;
+    int x = c.x;
+    int y = c.y;
     maze[y][x] = mode;
 }
