+++ Types and data structures/Lab2/generator.h
@@ -14,10 +14,10 @@ bool generate             (Data_t* d);
 
 bool generateStep         (Data_t* d);
 
-MazeMatrix_t init         (uint16_t width, uint16_t height);
+MazeMatrix_t init         (int width, int height);
 
-MazeMatrix_t finalize     (uint16_t width, uint16_t height, MazeMatrix_t maze);
+MazeMatrix_t finalize     (int width, int height, MazeMatrix_t maze);
 
-uint32_t unvisitedCount(uint16_t width, uint16_t height, MazeMatrix_t maze);
+int unvisitedCount(int width, int height, MazeMatrix_t maze);
 
 #endif // GENERATOR_H
